refactor: created ideation component
